Happy meal: Food Database API Search

Consumers with health or diet conditions need an easy way to get the recipe of foods they want to eat during the day, and the nutrient information of foods in their daily life. The function allows to search for food database matching the specified query, and bring health and happiness to your daily meal.

food_database_search(app_id_food, app_key_food, ingr)

Arguments

app_id_food
Application ID for food database API.

app_key_food
Food database API key.

ingr
The ingredient that the user wants for the food database. URL-encode should be included this string.

Details

The function provides you with tools to find nutrition and diet data for generic foods, packaged foods, and restaurant meals matching the specified queries, including app id, app key, ingredients, etc. Then the user can get data frame of food database matching the conditions.

Examples

app_key_food = Sys.getenv("food_api_key")
food_database_search(app_id = "bd9e5519", app_key_food = app_key_food, ingr = "large%20apple")

#>                   label name
#> 1                food.foodId
#> 2                   food.uri
#> 3                 food.label
#> 4  food.nutrients.ENERC_KCAL
#> 5      food.nutrients.PROCNT
#> 6         food.nutrients.FAT
#> 7      food.nutrients.CHOCDF
#> 8       food.nutrients.FIBTG
#> 9              food.category
#> 10        food.categoryLabel
#> 11                  quantity
#> 12               measure.uri
#> 13             measure.label
#>                                                       value
#> 1                         food_bnbh4ycaqj9as0a9z7h9xb2wmgat
#> 2    http://www.edamam.com/ontologies/edamam.owl#Food_09003
#> 3                                    Apples, raw, with skin
#> 4                                                        52
#> 5                                                      0.26
#> 6                                                      0.17
#> 7                                                     13.81
#> 8                                                       2.4
#> 9                                             Generic foods
#> 10                                                     food
#> 11                                                       20
#> 12 http://www.edamam.com/ontologies/edamam.owl#Measure_unit
#> 13                                                    Whole