Happy meal: Nutrients API Search

Consumers with health or diet conditions need an easy way to get the recipe of foods they want to eat during the day, and the nutrient information of foods in their daily life. The function allows to search for nutrient information of specified foods within several seconds, and bring health and happiness to your daily meal.

nutrient_search(app_id_nutrient, app_key_nutrient, ingr)

Arguments

app_id_nutrient
Application ID for nutrient API.

app_key_nutrient
Nutrient API key.

ingr
The ingredient that the user wants for the nutrient information. URL-encode should be included this string.

Details

The function allows to search for nutrient information of specified foods. It asks the user to insert specified queries, including app id, app key, ingredients, etc. Then the user can get data frame of the nutrient information matching the conditions.

Examples

app_key_nutrient = Sys.getenv("nutrient_api_key")
nutrient_search(app_id_nutrient = "a81f90d1",
app_key_nutrient = app_key_nutrient, ingr = "large%20apple")

#>                             label name
#> 1                                  uri
#> 2                             calories
#> 3                          totalWeight
#> 4                           dietLabels
#> 5                        healthLabels1
#> 6                        healthLabels2
#> 7                        healthLabels3
#> 8                        healthLabels4
#> 9                        healthLabels5
#> 10                            cautions
#> 11     totalNutrients.ENERC_KCAL.label
#> 12  totalNutrients.ENERC_KCAL.quantity
#> 13      totalNutrients.ENERC_KCAL.unit
#> 14            totalNutrients.FAT.label
#> 15         totalNutrients.FAT.quantity
#> 16             totalNutrients.FAT.unit
#> 17          totalNutrients.FASAT.label
#> 18       totalNutrients.FASAT.quantity
#> 19           totalNutrients.FASAT.unit
#> 20           totalNutrients.FAMS.label
#> 21        totalNutrients.FAMS.quantity
#> 22            totalNutrients.FAMS.unit
#> 23           totalNutrients.FAPU.label
#> 24        totalNutrients.FAPU.quantity
#> 25            totalNutrients.FAPU.unit
#> 26         totalNutrients.CHOCDF.label
#> 27      totalNutrients.CHOCDF.quantity
#> 28          totalNutrients.CHOCDF.unit
#> 29          totalNutrients.FIBTG.label
#> 30       totalNutrients.FIBTG.quantity
#> 31           totalNutrients.FIBTG.unit
#> 32          totalNutrients.SUGAR.label
#> 33       totalNutrients.SUGAR.quantity
#> 34           totalNutrients.SUGAR.unit
#> 35         totalNutrients.PROCNT.label
#> 36      totalNutrients.PROCNT.quantity
#> 37          totalNutrients.PROCNT.unit
#> 38             totalNutrients.NA.label
#> 39          totalNutrients.NA.quantity
#> 40              totalNutrients.NA.unit
#> 41             totalNutrients.CA.label
#> 42          totalNutrients.CA.quantity
#> 43              totalNutrients.CA.unit
#> 44             totalNutrients.MG.label
#> 45          totalNutrients.MG.quantity
#> 46              totalNutrients.MG.unit
#> 47              totalNutrients.K.label
#> 48           totalNutrients.K.quantity
#> 49               totalNutrients.K.unit
#> 50             totalNutrients.FE.label
#> 51          totalNutrients.FE.quantity
#> 52              totalNutrients.FE.unit
#> 53             totalNutrients.ZN.label
#> 54          totalNutrients.ZN.quantity
#> 55              totalNutrients.ZN.unit
#> 56              totalNutrients.P.label
#> 57           totalNutrients.P.quantity
#> 58               totalNutrients.P.unit
#> 59       totalNutrients.VITA_RAE.label
#> 60    totalNutrients.VITA_RAE.quantity
#> 61        totalNutrients.VITA_RAE.unit
#> 62           totalNutrients.VITC.label
#> 63        totalNutrients.VITC.quantity
#> 64            totalNutrients.VITC.unit
#> 65           totalNutrients.THIA.label
#> 66        totalNutrients.THIA.quantity
#> 67            totalNutrients.THIA.unit
#> 68           totalNutrients.RIBF.label
#> 69        totalNutrients.RIBF.quantity
#> 70            totalNutrients.RIBF.unit
#> 71            totalNutrients.NIA.label
#> 72         totalNutrients.NIA.quantity
#> 73             totalNutrients.NIA.unit
#> 74         totalNutrients.VITB6A.label
#> 75      totalNutrients.VITB6A.quantity
#> 76          totalNutrients.VITB6A.unit
#> 77         totalNutrients.FOLDFE.label
#> 78      totalNutrients.FOLDFE.quantity
#> 79          totalNutrients.FOLDFE.unit
#> 80          totalNutrients.FOLFD.label
#> 81       totalNutrients.FOLFD.quantity
#> 82           totalNutrients.FOLFD.unit
#> 83         totalNutrients.TOCPHA.label
#> 84      totalNutrients.TOCPHA.quantity
#> 85          totalNutrients.TOCPHA.unit
#> 86          totalNutrients.VITK1.label
#> 87       totalNutrients.VITK1.quantity
#> 88           totalNutrients.VITK1.unit
#> 89         totalDaily.ENERC_KCAL.label
#> 90      totalDaily.ENERC_KCAL.quantity
#> 91          totalDaily.ENERC_KCAL.unit
#> 92                totalDaily.FAT.label
#> 93             totalDaily.FAT.quantity
#> 94                 totalDaily.FAT.unit
#> 95              totalDaily.FASAT.label
#> 96           totalDaily.FASAT.quantity
#> 97               totalDaily.FASAT.unit
#> 98             totalDaily.CHOCDF.label
#> 99          totalDaily.CHOCDF.quantity
#> 100             totalDaily.CHOCDF.unit
#> 101             totalDaily.FIBTG.label
#> 102          totalDaily.FIBTG.quantity
#> 103              totalDaily.FIBTG.unit
#> 104            totalDaily.PROCNT.label
#> 105         totalDaily.PROCNT.quantity
#> 106             totalDaily.PROCNT.unit
#> 107                totalDaily.NA.label
#> 108             totalDaily.NA.quantity
#> 109                 totalDaily.NA.unit
#> 110                totalDaily.CA.label
#> 111             totalDaily.CA.quantity
#> 112                 totalDaily.CA.unit
#> 113                totalDaily.MG.label
#> 114             totalDaily.MG.quantity
#> 115                 totalDaily.MG.unit
#> 116                 totalDaily.K.label
#> 117              totalDaily.K.quantity
#> 118                  totalDaily.K.unit
#> 119                totalDaily.FE.label
#> 120             totalDaily.FE.quantity
#> 121                 totalDaily.FE.unit
#> 122                totalDaily.ZN.label
#> 123             totalDaily.ZN.quantity
#> 124                 totalDaily.ZN.unit
#> 125                 totalDaily.P.label
#> 126              totalDaily.P.quantity
#> 127                  totalDaily.P.unit
#> 128          totalDaily.VITA_RAE.label
#> 129       totalDaily.VITA_RAE.quantity
#> 130           totalDaily.VITA_RAE.unit
#> 131              totalDaily.VITC.label
#> 132           totalDaily.VITC.quantity
#> 133               totalDaily.VITC.unit
#> 134              totalDaily.THIA.label
#> 135           totalDaily.THIA.quantity
#> 136               totalDaily.THIA.unit
#> 137              totalDaily.RIBF.label
#> 138           totalDaily.RIBF.quantity
#> 139               totalDaily.RIBF.unit
#> 140               totalDaily.NIA.label
#> 141            totalDaily.NIA.quantity
#> 142                totalDaily.NIA.unit
#> 143            totalDaily.VITB6A.label
#> 144         totalDaily.VITB6A.quantity
#> 145             totalDaily.VITB6A.unit
#> 146            totalDaily.FOLDFE.label
#> 147         totalDaily.FOLDFE.quantity
#> 148             totalDaily.FOLDFE.unit
#> 149            totalDaily.TOCPHA.label
#> 150         totalDaily.TOCPHA.quantity
#> 151             totalDaily.TOCPHA.unit
#> 152             totalDaily.VITK1.label
#> 153          totalDaily.VITK1.quantity
#> 154              totalDaily.VITK1.unit
#>                                                                                   value
#> 1   http://www.edamam.com/ontologies/edamam.owl#recipe_4cac084761514da1a92fb1326b71e8a9
#> 2                                                                                  1892
#> 3                                                                                  3640
#> 4                                                                               LOW_FAT
#> 5                                                                                 VEGAN
#> 6                                                                            VEGETARIAN
#> 7                                                                           PEANUT_FREE
#> 8                                                                         TREE_NUT_FREE
#> 9                                                                          ALCOHOL_FREE
#> 10                                                                             SULFITES
#> 11                                                                               Energy
#> 12                                                                               1892.8
#> 13                                                                                 kcal
#> 14                                                                                  Fat
#> 15                                                                                6.188
#> 16                                                                                    g
#> 17                                                                            Saturated
#> 18                                                                               1.0192
#> 19                                                                                    g
#> 20                                                                      Monounsaturated
#> 21                                                                               0.2548
#> 22                                                                                    g
#> 23                                                                      Polyunsaturated
#> 24                                                                               1.8564
#> 25                                                                                    g
#> 26                                                                                Carbs
#> 27                                                                              502.684
#> 28                                                                                    g
#> 29                                                                                Fiber
#> 30                                                                                87.36
#> 31                                                                                    g
#> 32                                                                               Sugars
#> 33                                                                              378.196
#> 34                                                                                    g
#> 35                                                                              Protein
#> 36                                                                                9.464
#> 37                                                                                    g
#> 38                                                                               Sodium
#> 39                                                                                 36.4
#> 40                                                                                   mg
#> 41                                                                              Calcium
#> 42                                                                                218.4
#> 43                                                                                   mg
#> 44                                                                            Magnesium
#> 45                                                                                  182
#> 46                                                                                   mg
#> 47                                                                            Potassium
#> 48                                                                               3894.8
#> 49                                                                                   mg
#> 50                                                                                 Iron
#> 51                                                                                4.368
#> 52                                                                                   mg
#> 53                                                                                 Zinc
#> 54                                                                                1.456
#> 55                                                                                   mg
#> 56                                                                           Phosphorus
#> 57                                                                                400.4
#> 58                                                                                   mg
#> 59                                                                            Vitamin A
#> 60                                                                                109.2
#> 61                                                                                <b5>g
#> 62                                                                            Vitamin C
#> 63                                                                               167.44
#> 64                                                                                   mg
#> 65                                                                         Thiamin (B1)
#> 66                                                                               0.6188
#> 67                                                                                   mg
#> 68                                                                      Riboflavin (B2)
#> 69                                                                               0.9464
#> 70                                                                                   mg
#> 71                                                                          Niacin (B3)
#> 72                                                                               3.3124
#> 73                                                                                   mg
#> 74                                                                           Vitamin B6
#> 75                                                                               1.4924
#> 76                                                                                   mg
#> 77                                                            Folate equivalent (total)
#> 78                                                                                109.2
#> 79                                                                                <b5>g
#> 80                                                                        Folate (food)
#> 81                                                                                109.2
#> 82                                                                                <b5>g
#> 83                                                                            Vitamin E
#> 84                                                                                6.552
#> 85                                                                                   mg
#> 86                                                                            Vitamin K
#> 87                                                                                80.08
#> 88                                                                                <b5>g
#> 89                                                                               Energy
#> 90                                                                                94.64
#> 91                                                                                    %
#> 92                                                                                  Fat
#> 93                                                                                 9.52
#> 94                                                                                    %
#> 95                                                                            Saturated
#> 96                                                                                5.096
#> 97                                                                                    %
#> 98                                                                                Carbs
#> 99                                                                     167.561333333333
#> 100                                                                                   %
#> 101                                                                               Fiber
#> 102                                                                              349.44
#> 103                                                                                   %
#> 104                                                                             Protein
#> 105                                                                              18.928
#> 106                                                                                   %
#> 107                                                                              Sodium
#> 108                                                                    1.51666666666667
#> 109                                                                                   %
#> 110                                                                             Calcium
#> 111                                                                               21.84
#> 112                                                                                   %
#> 113                                                                           Magnesium
#> 114                                                                    43.3333333333333
#> 115                                                                                   %
#> 116                                                                           Potassium
#> 117                                                                     82.868085106383
#> 118                                                                                   %
#> 119                                                                                Iron
#> 120                                                                    24.2666666666667
#> 121                                                                                   %
#> 122                                                                                Zinc
#> 123                                                                    13.2363636363636
#> 124                                                                                   %
#> 125                                                                          Phosphorus
#> 126                                                                                57.2
#> 127                                                                                   %
#> 128                                                                           Vitamin A
#> 129                                                                    12.1333333333333
#> 130                                                                                   %
#> 131                                                                           Vitamin C
#> 132                                                                    186.044444444444
#> 133                                                                                   %
#> 134                                                                        Thiamin (B1)
#> 135                                                                    51.5666666666667
#> 136                                                                                   %
#> 137                                                                     Riboflavin (B2)
#> 138                                                                                72.8
#> 139                                                                                   %
#> 140                                                                         Niacin (B3)
#> 141                                                                             20.7025
#> 142                                                                                   %
#> 143                                                                          Vitamin B6
#> 144                                                                               114.8
#> 145                                                                                   %
#> 146                                                           Folate equivalent (total)
#> 147                                                                                27.3
#> 148                                                                                   %
#> 149                                                                           Vitamin E
#> 150                                                                               43.68
#> 151                                                                                   %
#> 152                                                                           Vitamin K
#> 153                                                                    66.7333333333333
#> 154                                                                                   %