Happy meal: Recipe API Search

Consumers with health or diet conditions need an easy way to get the recipe of foods they want to eat during the day, and the nutrient information of foods in their daily life. The function allows to search for recipes matching the specified query, and bring health and happiness to your daily meal.

recipes_search(q, app_id_recipe, app_key_recipe, diet, calories, excluded)

Arguments

q
The query text, what item you want the recipes focus on, for example, chicken.

app_id_recipe
Application ID for recipe API.

app_key_recipe
Recipe API key.

diet
Diet label: one of “balanced”, “high-protein”, “high-fiber”, “low-fat”, “low-carb”, “low-sodium”.

calories
Calories that each recipe serves, for example, “calories=100-300” will return all recipes with which have between 100 and 300 kcal per serving.

excluded
Excluding recipes with certain ingredients. More than one food can be excluded at the same time. Example: excluded=vinegar&excluded=pretzel，will exclude any recipes which contain vinegar or pretzels in their ingredient list.

Details

The function allows to search for recipes matching the specified query. It asks the user to pick search for recipes, and insert specified queries, including app id, app key, diet, calories, etc. Then the user can get data frame of the recipes matching the conditions.

Examples

app_key_recipe = Sys.getenv("recipe_api_key")
recipes_search(q = "pork", app_id_recipe = "46707447",
app_key_recipe = app_key_recipe, diet = "balanced",
calories = "100-1000", excluded = "onion")

#>                                                         label         source
#> 1                                       Yuzu-Cured Pork Belly   Zen Can Cook
#> 2                                          Pork & apple sauce   Jamie Oliver
#> 3                           Banana Peel Steamed Pork And Rice     No Recipes
#> 4                  Slow-Cooked Pulled Pork Sandwiches recipes     Epicurious
#> 5                                      Haw and Pork Ribs Soup       Red Cook
#> 6                                        Spicy Tofu with Pork         Food52
#> 7  Pork and Shrimp Siu Mai (Steamed Chinese Dumplings) Recipe   Serious Eats
#> 8                            Pork and Chive Dumplings recipes  Tasting Table
#> 9                                            Sticky Thai Pork  BBC Good Food
#> 10                                Pulled-Pork Sliders recipes Martha Stewart
#>                                                                                                     url
#> 1                                              http://www.zencancook.com/2010/03/yuzu-cured-pork-belly/
#> 2                                     http://www.jamieoliver.com/recipes/pork-recipes/pork-apple-sauce/
#> 3                               http://norecipes.com/blog/2008/04/25/banana-peel-steamed-pork-and-rice/
#> 4                http://www.epicurious.com/recipes/food/views/Slow-Cooked-Pulled-Pork-Sandwiches-365091
#> 5                                         http://redcook.net/2009/11/08/haw-fruit-treats-food-medicine/
#> 6                                                  https://food52.com/recipes/4475-spicy-tofu-with-pork
#> 7  http://www.seriouseats.com/recipes/2015/03/pork-shrimp-siu-mai-steamed-chinese-dumplings-recipe.html
#> 8     http://www.tastingtable.com/entry_detail/national/17617/recipes_home/Pork_and_Chive_Dumplings.htm
#> 9                                             http://www.bbcgoodfood.com/recipes/12463/sticky-thai-pork
#> 10                                              http://www.marthastewart.com/925950/pulled-pork-sliders
#>                                              healthLabels
#> 1                  Peanut-Free;Tree-Nut-Free;Alcohol-Free
#> 2                  Peanut-Free;Tree-Nut-Free;Alcohol-Free
#> 3                  Peanut-Free;Tree-Nut-Free;Alcohol-Free
#> 4                  Peanut-Free;Tree-Nut-Free;Alcohol-Free
#> 5               Sugar-Conscious;Peanut-Free;Tree-Nut-Free
#> 6              Sugar-Conscious;Tree-Nut-Free;Alcohol-Free
#> 7               Sugar-Conscious;Peanut-Free;Tree-Nut-Free
#> 8  Sugar-Conscious;Peanut-Free;Tree-Nut-Free;Alcohol-Free
#> 9                              Tree-Nut-Free;Alcohol-Free
#> 10                 Peanut-Free;Tree-Nut-Free;Alcohol-Free
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               ingredients
#> 1                                                                                                                                                                                                                                                       2 pieces of 2 lbs pork belly;1/3 cup kosher salt;1/3 cup sugar;Grated zest from 4 yuzu or 3 lemons;1 tablespoon rosemary, finely chopped;1/2 lb fresh peas, blanched;1/2 lb fresh fava beans, blanched;Shredded butter head lettuce;3 tablespoons butter;Chicken stock;Salt, to taste;Hot pepper powder, to taste;Vinegar reduction, to taste;56.6;77.6666666666667;66.6666666666667;232;1.7;226.796185;226.796185;12.6592295653333;42.6;186.165140666667;5.58495422;27.9247711;0
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                           2 green eating apples;1cm piece of ginger;150 g brown rice;200 g small carrots;200 g tenderstem broccoli;220 g piece of lean higher-welfare pork fillet;1 whole nutmeg , for grating;8 leaves of fresh sage;olive oil;2 oranges;2 tablespoons natural yogurt;364;28.688;150;200;200;220;2;8;19.50784;262;28.4
#> 3                                                                                                                                                                                                                                                                                                                                    1/2 C sour orange juice (or 1/4 C orange juice + 1/4 C lemon juice);1 Tbs olive oil;2 cloves garlic pressed;1 1/2 tsp oregano;minced fresh cilantro;1 tsp kosher salt;fresh ground black pepper;1 lbs boneless country style pork ribs (or other well marbled cut);1 cup long grain rice;5 large organic bananas;124;13.5;6;1.5;14.6844653666691;4.85416666691288;4.40533961000074;453.59237;185;680
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           1 3 1/2-pound boneless pork shoulder or butt roast, rind and excess fat trimmed;ground black pepper;salt;2 18-22-ounce bottles of your favorite barbecue sauce;8 sesame seed buns, toasted;1587.573295;9.17266266;18.34532532;1133.980925;336
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                             12 oz. pork ribs;8 oz. fresh haw fruit (or 4 oz. dried haw);4 oz. prunes;1 ear fresh corn cut into 1-inch pieces;4 slices ginger root;1/4 cup shaoxing cooking wine;6 cups chicken stock or water;1 teaspoon salt;1/2 teaspoon ground white pepper;340.1942775;226.796185;113.3980925;102;8.8;58;1440;6;1.2
#> 6                                                                                                                                                                                                                                                                                                                         1.5 tablespoons peanut oil;1/3 pound ground pork;1 tablespoon garlic;2 teaspoons ginger;2 scallions;1 packet firm silken tofu;1 tablespoon chili garlic paste;3 teaspoons tamari;1 teaspoon sesame oil;1 teaspoon white sugar;1 teaspoon corn starch;2 tablespoons cold water;1 cup jasmine rice;20.25;151.197456666667;8.49999999985629;4;30;324;8.49999999985629;18;4.5;4.2;3.02083333348655;29.573529562;185
#> 7  1/4 pound shrimp, shelled and deveined;1 tablespoon baking soda;1/2 pound bonesless pork, such as shoulder, cut into large cubes;2 ounces pork fat, such as fatback or fatty belly, finely diced;1/4 teaspoon ground white pepper;1/2 teaspoon kosher salt;2 teaspoons cornstarch;2 teaspoons Shaoxing wine;1 teaspoon toasted sesame oil;1 teaspoon extra-virgin olive oil;1/2 teaspoon finely grated fresh peeled ginger;1/2 teaspoon sugar;1 package extra-thin wonton wrappers or regular wonton wrappers;1 large carrot, thinly sliced into 15-20 rounds and the rest finely minced (optional, see note above);113.3980925;13.7999999990667;226.796185;56.69904625;0.6;2.42708333345644;5.2;9.83333333383209;4.5;4.5;1;2.1;448;72
#> 8                                                                                                                                                                                                                                                                                                                                                                   2 teaspoons toasted white sesame seeds;1 pound ground pork;1 cup finely chopped garlic chives;2 teaspoon minced garlic;1¼ teaspoon salt;1 teaspoon cornstarch;1 teaspoon grated ginger;1 teaspoon freshly ground black pepper;Thirty 3½-inch round dumpling wrappers (see note above);canola oil;6.00000000030433;453.59237;48;5.6;7.5;2.6;2;2.9;960;20.2394162320041
#> 9                                                                                                                                                                                                                100.0g palm sugar or light muscovado sugar;1 thumb-size piece fresh root ginger;200.0g minced pork;½ pineapple , peeled and sliced very thinly, to serve;Small bunch coriander , stalks finely chopped (or if you can find a bunch with the root intact, chop that instead);1.0 tsp fish sauce;2 fat green chillies , seeds left in;1 garlic clove , roughly chopped;1.0 tbsp groundnut or vegetable oil;50.0g dried shrimps , soaked in water for 10 mins, then drained;100;15;200;452.5;21;6.00000000040577;90;3;14;50
#> 10                                                                                                                                                                                                    1 bone-in pork shoulder (8 to 10 pounds), skin on;4 ounces slab bacon, cut into 1-inch pieces;freshly ground pepper;coarse salt;1/4 cup plus 1 tablespoon packed light-brown sugar;1/4 cup plus 1 tablespoon packed light-brown sugar;1/4 cup plus 1 tablespoon unsulfured molasses;1/4 cup plus 1 tablespoon unsulfured molasses;1/4 cup ketchup;3 tablespoons white vinegar;About 4 cups water;60 mini brioche buns, split;4082.33133;113.3980925;31.0242882674993;62.0485765349986;55;13.7499999997675;84.25;20;60;44.7;948;4920
#>     calories totalNutrients.ENERC_KCAL.label totalNutrients.ENERC_KCAL.quantity
#> 1   1297.259                          Energy                           1297.259
#> 2   1689.444                          Energy                           1689.444
#> 3   3111.336                          Energy                           3111.336
#> 4   5717.555                          Energy                           5717.555
#> 5   2040.509                          Energy                           2040.509
#> 6   1586.271                          Energy                           1586.271
#> 7   2339.344                          Energy                           2339.344
#> 8   4241.373                          Energy                           4241.373
#> 9   1285.660                          Energy                           1285.660
#> 10 25808.687                          Energy                          25808.687
#>    totalNutrients.ENERC_KCAL.unit totalNutrients.FAT.label
#> 1                            kcal                      Fat
#> 2                            kcal                      Fat
#> 3                            kcal                      Fat
#> 4                            kcal                      Fat
#> 5                            kcal                      Fat
#> 6                            kcal                      Fat
#> 7                            kcal                      Fat
#> 8                            kcal                      Fat
#> 9                            kcal                      Fat
#> 10                           kcal                      Fat
#>    totalNutrients.FAT.quantity totalNutrients.FAT.unit
#> 1                     57.65631                       g
#> 2                     56.24140                       g
#> 3                    151.15549                       g
#> 4                    234.28004                       g
#> 5                     99.34046                       g
#> 6                     71.75132                       g
#> 7                     82.84317                       g
#> 8                    134.22508                       g
#> 9                     44.46380                       g
#> 10                  1076.99202                       g
#>    totalNutrients.FASAT.label totalNutrients.FASAT.quantity
#> 1                   Saturated                      32.03001
#> 2                   Saturated                      15.11756
#> 3                   Saturated                      51.56491
#> 4                   Saturated                      77.91001
#> 5                   Saturated                      30.67982
#> 6                   Saturated                      19.13426
#> 7                   Saturated                      25.11766
#> 8                   Saturated                      40.25030
#> 9                   Saturated                      10.08177
#> 10                  Saturated                     266.58894
#>    totalNutrients.FASAT.unit totalNutrients.FATRN.label
#> 1                          g                      Trans
#> 2                          g                       <NA>
#> 3                          g                      Trans
#> 4                          g                       <NA>
#> 5                          g                      Trans
#> 6                          g                      Trans
#> 7                          g                      Trans
#> 8                          g                      Trans
#> 9                          g                      Trans
#> 10                         g                      Trans
#>    totalNutrients.FATRN.quantity totalNutrients.FATRN.unit
#> 1                   1.8113975452                         g
#> 2                             NA                      <NA>
#> 3                   0.2177243376                         g
#> 4                             NA                      <NA>
#> 5                   0.7623712961                         g
#> 6                   0.0002114583                         g
#> 7                   0.3606059342                         g
#> 8                   0.0799456941                         g
#> 9                   0.1163800000                         g
#> 10                  0.1508194630                         g
#>    totalNutrients.FAMS.label totalNutrients.FAMS.quantity
#> 1            Monounsaturated                     16.24197
#> 2            Monounsaturated                     28.67915
#> 3            Monounsaturated                     68.75063
#> 4            Monounsaturated                     99.39848
#> 5            Monounsaturated                     38.48105
#> 6            Monounsaturated                     29.80062
#> 7            Monounsaturated                     34.78803
#> 8            Monounsaturated                     58.69450
#> 9            Monounsaturated                     24.34502
#> 10           Monounsaturated                    265.29318
#>    totalNutrients.FAMS.unit totalNutrients.FAPU.label
#> 1                         g           Polyunsaturated
#> 2                         g           Polyunsaturated
#> 3                         g           Polyunsaturated
#> 4                         g           Polyunsaturated
#> 5                         g           Polyunsaturated
#> 6                         g           Polyunsaturated
#> 7                         g           Polyunsaturated
#> 8                         g           Polyunsaturated
#> 9                         g           Polyunsaturated
#> 10                        g           Polyunsaturated
#>    totalNutrients.FAPU.quantity totalNutrients.FAPU.unit
#> 1                      4.324799                        g
#> 2                      7.218558                        g
#> 3                     15.860134                        g
#> 4                     29.453759                        g
#> 5                     17.276570                        g
#> 6                     17.477703                        g
#> 7                     16.023540                        g
#> 8                     21.696547                        g
#> 9                      6.015570                        g
#> 10                    66.490631                        g
#>    totalNutrients.CHOCDF.label totalNutrients.CHOCDF.quantity
#> 1                        Carbs                       170.4488
#> 2                        Carbs                       240.0139
#> 3                        Carbs                       322.4890
#> 4                        Carbs                       636.6939
#> 5                        Carbs                       194.7996
#> 6                        Carbs                       167.6828
#> 7                        Carbs                       275.0240
#> 8                        Carbs                       565.7693
#> 9                        Carbs                       183.1556
#> 10                       Carbs                      2883.0948
#>    totalNutrients.CHOCDF.unit totalNutrients.FIBTG.label
#> 1                           g                      Fiber
#> 2                           g                      Fiber
#> 3                           g                      Fiber
#> 4                           g                      Fiber
#> 5                           g                      Fiber
#> 6                           g                      Fiber
#> 7                           g                      Fiber
#> 8                           g                      Fiber
#> 9                           g                      Fiber
#> 10                          g                      Fiber
#>    totalNutrients.FIBTG.quantity totalNutrients.FIBTG.unit
#> 1                      35.730891                         g
#> 2                      35.137760                         g
#> 3                      22.622216                         g
#> 4                      19.582512                         g
#> 5                       4.798362                         g
#> 6                       6.742417                         g
#> 7                      10.304000                         g
#> 8                      20.102700                         g
#> 9                      16.847000                         g
#> 10                     68.028545                         g
#>    totalNutrients.SUGAR.label totalNutrients.SUGAR.quantity
#> 1                      Sugars                    110.535183
#> 2                      Sugars                     77.204336
#> 3                      Sugars                     94.079299
#> 4                      Sugars                    398.397165
#> 5                      Sugars                     30.185800
#> 6                      Sugars                      7.789704
#> 7                      Sugars                      5.603283
#> 8                      Sugars                      1.014560
#> 9                      Sugars                    149.644650
#> 10                     Sugars                    638.709386
#>    totalNutrients.SUGAR.unit totalNutrients.SUGAR.added.label
#> 1                          g                    Sugars, added
#> 2                          g                             <NA>
#> 3                          g                             <NA>
#> 4                          g                             <NA>
#> 5                          g                             <NA>
#> 6                          g                    Sugars, added
#> 7                          g                    Sugars, added
#> 8                          g                             <NA>
#> 9                          g                    Sugars, added
#> 10                         g                    Sugars, added
#>    totalNutrients.SUGAR.added.quantity totalNutrients.SUGAR.added.unit
#> 1                             66.53333                               g
#> 2                                   NA                            <NA>
#> 3                                   NA                            <NA>
#> 4                                   NA                            <NA>
#> 5                                   NA                            <NA>
#> 6                              4.19160                               g
#> 7                              2.09580                               g
#> 8                                   NA                            <NA>
#> 9                             99.80000                               g
#> 10                           144.59685                               g
#>    totalNutrients.PROCNT.label totalNutrients.PROCNT.quantity
#> 1                      Protein                       49.71898
#> 2                      Protein                       68.06200
#> 3                      Protein                      121.41393
#> 4                      Protein                      247.97370
#> 5                      Protein                       97.81593
#> 6                      Protein                       68.94181
#> 7                      Protein                      110.12231
#> 8                      Protein                      173.98042
#> 9                      Protein                       53.77660
#> 10                     Protein                     1024.24927
#>    totalNutrients.PROCNT.unit totalNutrients.CHOLE.label
#> 1                           g                Cholesterol
#> 2                           g                Cholesterol
#> 3                           g                Cholesterol
#> 4                           g                Cholesterol
#> 5                           g                Cholesterol
#> 6                           g                Cholesterol
#> 7                           g                Cholesterol
#> 8                           g                Cholesterol
#> 9                           g                Cholesterol
#> 10                          g                Cholesterol
#>    totalNutrients.CHOLE.quantity totalNutrients.CHOLE.unit
#> 1                       160.0503                        mg
#> 2                       142.2920                        mg
#> 3                       412.7691                        mg
#> 4                       845.3828                        mg
#> 5                       315.3554                        mg
#> 6                       108.8622                        mg
#> 7                       366.9065                        mg
#> 8                       412.9865                        mg
#> 9                       189.0000                        mg
#> 10                     4948.6867                        mg
#>    totalNutrients.NA.label totalNutrients.NA.quantity totalNutrients.NA.unit
#> 1                   Sodium                  2511.3049                     mg
#> 2                   Sodium                   342.0236                     mg
#> 3                   Sodium                  2143.8369                     mg
#> 4                   Sodium                 14101.7606                     mg
#> 5                   Sodium                  5052.2952                     mg
#> 6                   Sodium                  1148.1686                     mg
#> 7                   Sodium                  7171.0305                     mg
#> 8                   Sodium                  5749.3377                     mg
#> 9                   Sodium                   877.4950                     mg
#> 10                  Sodium                 33987.8723                     mg
#>    totalNutrients.CA.label totalNutrients.CA.quantity totalNutrients.CA.unit
#> 1                  Calcium                   241.2556                     mg
#> 2                  Calcium                   550.7292                     mg
#> 3                  Calcium                   278.3073                     mg
#> 4                  Calcium                  1174.7306                     mg
#> 5                  Calcium                   205.0395                     mg
#> 6                  Calcium                   781.8073                     mg
#> 7                  Calcium                   347.0228                     mg
#> 8                  Calcium                   640.7179                     mg
#> 9                  Calcium                   294.7250                     mg
#> 10                 Calcium                  2113.2646                     mg
#>    totalNutrients.MG.label totalNutrients.MG.quantity totalNutrients.MG.unit
#> 1                Magnesium                   196.7080                     mg
#> 2                Magnesium                   424.7438                     mg
#> 3                Magnesium                   360.2300                     mg
#> 4                Magnesium                   458.0652                     mg
#> 5                Magnesium                   244.4097                     mg
#> 6                Magnesium                   215.4410                     mg
#> 7                Magnesium                   176.4244                     mg
#> 8                Magnesium                   326.6996                     mg
#> 9                Magnesium                   215.0000                     mg
#> 10               Magnesium                   893.9735                     mg
#>    totalNutrients.K.label totalNutrients.K.quantity totalNutrients.K.unit
#> 1               Potassium                  2139.785                    mg
#> 2               Potassium                  3576.770                    mg
#> 3               Potassium                  4534.254                    mg
#> 4               Potassium                  6775.314                    mg
#> 5               Potassium                  4076.239                    mg
#> 6               Potassium                  1340.177                    mg
#> 7               Potassium                  1736.668                    mg
#> 8               Potassium                  2328.545                    mg
#> 9               Potassium                  1911.155                    mg
#> 10              Potassium                 11670.473                    mg
#>    totalNutrients.FE.label totalNutrients.FE.quantity totalNutrients.FE.unit
#> 1                     Iron                   9.542565                     mg
#> 2                     Iron                   9.802772                     mg
#> 3                     Iron                   9.465216                     mg
#> 4                     Iron                  32.175083                     mg
#> 5                     Iron                  11.798489                     mg
#> 6                     Iron                  15.729721                     mg
#> 7                     Iron                  17.624977                     mg
#> 8                     Iron                  38.289623                     mg
#> 9                     Iron                   7.589250                     mg
#> 10                    Iron                 203.258725                     mg
#>    totalNutrients.ZN.label totalNutrients.ZN.quantity totalNutrients.ZN.unit
#> 1                     Zinc                   6.615227                     mg
#> 2                     Zinc                   9.171099                     mg
#> 3                     Zinc                  13.318182                     mg
#> 4                     Zinc                  37.074881                     mg
#> 5                     Zinc                  12.117860                     mg
#> 6                     Zinc                   8.456542                     mg
#> 7                     Zinc                   8.766303                     mg
#> 8                     Zinc                  17.636662                     mg
#> 9                     Zinc                   5.836800                     mg
#> 10                    Zinc                  84.898716                     mg
#>    totalNutrients.P.label totalNutrients.P.quantity totalNutrients.P.unit
#> 1              Phosphorus                  762.7854                    mg
#> 2              Phosphorus                 1156.3939                    mg
#> 3              Phosphorus                 1379.5267                    mg
#> 4              Phosphorus                 2767.8465                    mg
#> 5              Phosphorus                 1120.4755                    mg
#> 6              Phosphorus                  933.9441                    mg
#> 7              Phosphorus                 1163.2257                    mg
#> 8              Phosphorus                 1651.1346                    mg
#> 9              Phosphorus                  686.9000                    mg
#> 10             Phosphorus                 5837.1494                    mg
#>    totalNutrients.VITA_RAE.label totalNutrients.VITA_RAE.quantity
#> 1                      Vitamin A                        523.91057
#> 2                      Vitamin A                       1807.50800
#> 3                      Vitamin A                         93.82294
#> 4                      Vitamin A                        151.02812
#> 5                      Vitamin A                        173.26548
#> 6                      Vitamin A                         18.29582
#> 7                      Vitamin A                        699.63265
#> 8                      Vitamin A                        152.89485
#> 9                      Vitamin A                         88.01500
#> 10                     Vitamin A                         97.68532
#>    totalNutrients.VITA_RAE.unit totalNutrients.VITC.label
#> 1                            µg                 Vitamin C
#> 2                            µg                 Vitamin C
#> 3                            µg                 Vitamin C
#> 4                            µg                 Vitamin C
#> 5                            µg                 Vitamin C
#> 6                            µg                 Vitamin C
#> 7                            µg                 Vitamin C
#> 8                            µg                 Vitamin C
#> 9                            µg                 Vitamin C
#> 10                           µg                 Vitamin C
#>    totalNutrients.VITC.quantity totalNutrients.VITC.unit
#> 1                    263.280359                       mg
#> 2                    351.876400                       mg
#> 3                    128.392083                       mg
#> 4                     19.506645                       mg
#> 5                     14.817128                       mg
#> 6                     13.055799                       mg
#> 7                      5.784777                       mg
#> 8                     32.910347                       mg
#> 9                    352.951000                       mg
#> 10                    23.892239                       mg
#>    totalNutrients.THIA.label totalNutrients.THIA.quantity
#> 1               Thiamin (B1)                     1.596321
#> 2               Thiamin (B1)                     3.248168
#> 3               Thiamin (B1)                     4.200609
#> 4               Thiamin (B1)                    11.549018
#> 5               Thiamin (B1)                     1.917613
#> 6               Thiamin (B1)                     2.433568
#> 7               Thiamin (B1)                     4.530345
#> 8               Thiamin (B1)                     8.402428
#> 9               Thiamin (B1)                     2.294935
#> 10              Thiamin (B1)                    23.879438
#>    totalNutrients.THIA.unit totalNutrients.RIBF.label
#> 1                        mg           Riboflavin (B2)
#> 2                        mg           Riboflavin (B2)
#> 3                        mg           Riboflavin (B2)
#> 4                        mg           Riboflavin (B2)
#> 5                        mg           Riboflavin (B2)
#> 6                        mg           Riboflavin (B2)
#> 7                        mg           Riboflavin (B2)
#> 8                        mg           Riboflavin (B2)
#> 9                        mg           Riboflavin (B2)
#> 10                       mg           Riboflavin (B2)
#>    totalNutrients.RIBF.quantity totalNutrients.RIBF.unit
#> 1                     1.3601424                       mg
#> 2                     1.2376419                       mg
#> 3                     2.1628073                       mg
#> 4                     4.6189104                       mg
#> 5                     2.3790257                       mg
#> 6                     0.7239635                       mg
#> 7                     2.3694501                       mg
#> 8                     4.7768221                       mg
#> 9                     0.8174200                       mg
#> 10                    8.6691895                       mg
#>    totalNutrients.NIA.label totalNutrients.NIA.quantity totalNutrients.NIA.unit
#> 1               Niacin (B3)                    15.99992                      mg
#> 2               Niacin (B3)                    21.57212                      mg
#> 3               Niacin (B3)                    27.33627                      mg
#> 4               Niacin (B3)                    68.52387                      mg
#> 5               Niacin (B3)                    44.84589                      mg
#> 6               Niacin (B3)                    15.71253                      mg
#> 7               Niacin (B3)                    38.92704                      mg
#> 8               Niacin (B3)                    72.41604                      mg
#> 9               Niacin (B3)                    14.15068                      mg
#> 10              Niacin (B3)                   124.17785                      mg
#>    totalNutrients.VITB6A.label totalNutrients.VITB6A.quantity
#> 1                   Vitamin B6                       1.334135
#> 2                   Vitamin B6                       3.007729
#> 3                   Vitamin B6                       4.873611
#> 4                   Vitamin B6                       5.259304
#> 5                   Vitamin B6                       3.870246
#> 6                   Vitamin B6                       1.385986
#> 7                   Vitamin B6                       1.575399
#> 8                   Vitamin B6                       2.219698
#> 9                   Vitamin B6                       2.071510
#> 10                  Vitamin B6                      11.868267
#>    totalNutrients.VITB6A.unit totalNutrients.FOLDFE.label
#> 1                          mg   Folate equivalent (total)
#> 2                          mg   Folate equivalent (total)
#> 3                          mg   Folate equivalent (total)
#> 4                          mg   Folate equivalent (total)
#> 5                          mg   Folate equivalent (total)
#> 6                          mg   Folate equivalent (total)
#> 7                          mg   Folate equivalent (total)
#> 8                          mg   Folate equivalent (total)
#> 9                          mg   Folate equivalent (total)
#> 10                         mg   Folate equivalent (total)
#>    totalNutrients.FOLDFE.quantity totalNutrients.FOLDFE.unit
#> 1                        528.4036                         µg
#> 2                        314.3037                         µg
#> 3                        219.7320                         µg
#> 4                        658.3330                         µg
#> 5                        125.5798                         µg
#> 6                        809.7286                         µg
#> 7                        638.0819                         µg
#> 8                       1366.1806                         µg
#> 9                        118.4500                         µg
#> 10                       164.4491                         µg
#>    totalNutrients.FOLFD.label totalNutrients.FOLFD.quantity
#> 1               Folate (food)                      528.4036
#> 2               Folate (food)                      314.3037
#> 3               Folate (food)                      219.7320
#> 4               Folate (food)                      167.7730
#> 5               Folate (food)                      125.5798
#> 6               Folate (food)                      108.5786
#> 7               Folate (food)                      113.9219
#> 8               Folate (food)                      242.9806
#> 9               Folate (food)                      118.4500
#> 10              Folate (food)                      164.4491
#>    totalNutrients.FOLFD.unit totalNutrients.VITB12.label
#> 1                         µg                 Vitamin B12
#> 2                         µg                 Vitamin B12
#> 3                         µg                 Vitamin B12
#> 4                         µg                 Vitamin B12
#> 5                         µg                 Vitamin B12
#> 6                         µg                 Vitamin B12
#> 7                         µg                 Vitamin B12
#> 8                         µg                 Vitamin B12
#> 9                         µg                 Vitamin B12
#> 10                        µg                 Vitamin B12
#>    totalNutrients.VITB12.quantity totalNutrients.VITB12.unit
#> 1                       0.3939207                         µg
#> 2                       1.2710800                         µg
#> 3                       3.4473020                         µg
#> 4                       9.4830318                         µg
#> 5                       1.2927383                         µg
#> 6                       1.0583822                         µg
#> 7                       2.9302222                         µg
#> 8                       3.3671466                         µg
#> 9                       1.6438000                         µg
#> 10                     23.2239293                         µg
#>    totalNutrients.VITD.label totalNutrients.VITD.quantity
#> 1                  Vitamin D                     1.111888
#> 2                  Vitamin D                     1.128400
#> 3                  Vitamin D                     8.164663
#> 4                  Vitamin D                    20.241560
#> 5                  Vitamin D                     7.824468
#> 6                       <NA>                           NA
#> 7                  Vitamin D                     2.211263
#> 8                       <NA>                           NA
#> 9                  Vitamin D                     1.050000
#> 10                 Vitamin D                    52.503317
#>    totalNutrients.VITD.unit totalNutrients.TOCPHA.label
#> 1                        µg                   Vitamin E
#> 2                        µg                   Vitamin E
#> 3                        µg                   Vitamin E
#> 4                        µg                   Vitamin E
#> 5                        µg                   Vitamin E
#> 6                      <NA>                   Vitamin E
#> 7                        µg                   Vitamin E
#> 8                      <NA>                   Vitamin E
#> 9                        µg                   Vitamin E
#> 10                       µg                   Vitamin E
#>    totalNutrients.TOCPHA.quantity totalNutrients.TOCPHA.unit
#> 1                        4.923075                         mg
#> 2                        7.958204                         mg
#> 3                        3.561977                         mg
#> 4                       11.664735                         mg
#> 5                        1.784999                         mg
#> 6                        3.667240                         mg
#> 7                        3.397813                         mg
#> 8                        3.689442                         mg
#> 9                        4.884900                         mg
#> 10                       7.503587                         mg
#>    totalNutrients.VITK1.label totalNutrients.VITK1.quantity
#> 1                   Vitamin K                     157.15504
#> 2                   Vitamin K                     386.59721
#> 3                   Vitamin K                      73.99688
#> 4                   Vitamin K                      45.84331
#> 5                   Vitamin K                       3.19480
#> 6                   Vitamin K                      71.11681
#> 7                   Vitamin K                      13.16619
#> 8                   Vitamin K                     121.37120
#> 9                   Vitamin K                      15.98350
#> 10                  Vitamin K                      52.58676
#>    totalNutrients.VITK1.unit totalNutrients.FOLAC.label
#> 1                         µg                       <NA>
#> 2                         µg                       <NA>
#> 3                         µg                       <NA>
#> 4                         µg                 Folic acid
#> 5                         µg                       <NA>
#> 6                         µg                 Folic acid
#> 7                         µg                 Folic acid
#> 8                         µg                 Folic acid
#> 9                         µg                       <NA>
#> 10                        µg                       <NA>
#>    totalNutrients.FOLAC.quantity totalNutrients.FOLAC.unit
#> 1                             NA                      <NA>
#> 2                             NA                      <NA>
#> 3                             NA                      <NA>
#> 4                         288.96                        µg
#> 5                             NA                      <NA>
#> 6                         412.55                        µg
#> 7                         309.12                        µg
#> 8                         662.40                        µg
#> 9                             NA                      <NA>
#> 10                            NA                      <NA>