Happy meal: UPC or Barcode Search

Consumers with health or diet conditions need an easy way to get the recipe of foods they want to eat during the day, and the nutrient information of foods in their daily life. The function allows for search of an item based UPC/Barcode number, and let you know more about what you eat.

upc_barcode_search(upc, app_id_food, app_key_food)

Arguments

upc
The UPC or barcode number for the food. It is usually a 12-units number.

app_id_food
Application ID for food database API.

app_key_food
Food database API key.

Details

This is a service allowing you to submit a UPC or barcode and find a match for it in the food database. Then the user can get data frame of the detailed information matching the conditions.

Examples

app_key_food = Sys.getenv("food_api_key")
upc_barcode_search(upc = "041250102410", app_id_food = "bd9e5519", app_key_food = app_key_food)

#>                         label name
#> 1                             text
#> 2                hints.food.foodId
#> 3                   hints.food.uri
#> 4                 hints.food.label
#> 5  hints.food.nutrients.ENERC_KCAL
#> 6      hints.food.nutrients.PROCNT
#> 7         hints.food.nutrients.FAT
#> 8      hints.food.nutrients.CHOCDF
#> 9                 hints.food.brand
#> 10             hints.food.category
#> 11        hints.food.categoryLabel
#> 12    hints.food.foodContentsLabel
#> 13             hints.measures.uri1
#> 14             hints.measures.uri2
#> 15             hints.measures.uri3
#> 16             hints.measures.uri4
#> 17           hints.measures.label1
#> 18           hints.measures.label2
#> 19           hints.measures.label3
#> 20           hints.measures.label4
#>                                                                                                                                              value
#> 1                                                                                                                                 upc:041250102410
#> 2                                                                                                                food_bnfnm0nbls36lca43qmvvanadk0o
#> 3                                                                                 http://www.edamam.com/ontologies/edamam.owl#Food_USDABR_45241819
#> 4                                                                                                                       CULTURED LOWFAT BUTTERMILK
#> 5                                                                                                                                               46
#> 6                                                                                                                                 3.32999992370605
#> 7                                                                                                                                 1.03999996185303
#> 8                                                                                                                                 5.42000007629395
#> 9                                                                                                                                     Meijer, Inc.
#> 10                                                                                                                                  Packaged foods
#> 11                                                                                                                                            food
#> 12  CULTURED LOWFAT MILK; NONFAT MILK; MODIFIED FOOD STARCH; SALT; LOCUST BEAN GUM; SODIUM CITRATE; CARRAGEENAN; VITAMIN A PALMITATE; VITAMIN D3. 
#> 13                                                                                        http://www.edamam.com/ontologies/edamam.owl#Measure_gram
#> 14                                                                                       http://www.edamam.com/ontologies/edamam.owl#Measure_ounce
#> 15                                                                                       http://www.edamam.com/ontologies/edamam.owl#Measure_pound
#> 16                                                                                    http://www.edamam.com/ontologies/edamam.owl#Measure_kilogram
#> 17                                                                                                                                            Gram
#> 18                                                                                                                                           Ounce
#> 19                                                                                                                                           Pound
#> 20                                                                                                                                        Kilogram